debug

+++ train_mm.py
@@ -35,7 +35,7 @@ def batchPadding(batch):
         if total_video_len > max_video_len:
             max_video_len = total_video_len
     
-    text_tensor = torch.zeros(batch_size, max_text_len)
+    text_tensor = torch.zeros(batch_size, max_text_len, dtype=torch.long)
     video_tensor = torch.zeros(batch_size, max_video_len, video.shape[1], dtype=torch.float)
 
     segments_tensor = torch.cat([torch.zeros(batch_size, max_text_len, dtype=torch.long), torch.ones(batch_size, max_video_len, dtype=torch.long)], dim=1)
